Refactor code for improved readability
The changes in this commit mainly focus on reformatting and restructuring the code to improve readability. Such modifications are evident in changes to the path definitions in urls.py files, as well as some logical checks in views.py. Additionally, the redundant send_mail function was removed from the CommentCreateView in blog/views.py.