Update initial migration and various cleanups
The initial migration file for the blog has been updated with a new date. In addition, minor code cleanups were performed in the utils.py, forms.py,model.py, constants.py and validators.py files. The minor changes include type checking annotation modification, the removal of an unused constant, addition of a verbose name to a foreign key, and the use of newly defined Config parameters in validator logic.